+++ packages/codux-ui/src/image/index.ts
@@ -1,0 +1,2 @@
+export * from './image';
+export * from './types';
